test(model) : 기능 테스트 코드 수정
기존 Cars 클래스를 Car 클래스와 CarFactory 클래스로 분할하였다. 이에 따라 올바른 테스트를 위해 해당 코드를 그에 맞게 수정하였다.
